split zones layer and move all styles to top
