+++ web/src/features/map/map-layers/ZonesLayer.tsx
@@ -1,0 +1,53 @@
+import { useTheme } from 'hooks/theme';
+import { useMemo } from 'react';
+import { Layer, Source } from 'react-map-gl';
+
+import { ZONE_SOURCE } from '../Map';
+import { useGetGeometries } from '../map-utils/getMapGrid';
+
+export default function ZonesLayer() {
+  const { worldGeometries } = useGetGeometries();
+  const theme = useTheme();
+
+  const mapStyles = useMemo(
+    () => ({
+      zonesBorder: {
+        'line-color': [
+          'case',
+          ['boolean', ['feature-state', 'selected'], false],
+          'white',
+          theme.strokeColor,
+        ],
+        'line-width': [
+          'case',
+          ['boolean', ['feature-state', 'selected'], false],
+          theme.strokeWidth * 10,
+          ['boolean', ['feature-state', 'hover'], false],
+          theme.strokeWidth * 10,
+          theme.strokeWidth,
+        ],
+      } as mapboxgl.LinePaint,
+      zonesClickable: {
+        'fill-color': [
+          'coalesce',
+          ['feature-state', 'color'],
+          ['get', 'color'],
+          theme.clickableFill,
+        ],
+      } as mapboxgl.FillPaint,
+      zonesHover: {
+        'fill-color': '#FFFFFF',
+        'fill-opacity': ['case', ['boolean', ['feature-state', 'hover'], false], 0.3, 0],
+      } as mapboxgl.FillPaint,
+    }),
+    [theme]
+  );
+
+  return (
+    <Source id={ZONE_SOURCE} promoteId={'zoneId'} type="geojson" data={worldGeometries}>
+      <Layer id="zones-clickable-layer" type="fill" paint={mapStyles.zonesClickable} />
+      <Layer id="zones-hoverable-layer" type="fill" paint={mapStyles.zonesHover} />
+      <Layer id="zones-border" type="line" paint={mapStyles.zonesBorder} />
+    </Source>
+  );
+}
